+++ Source/lua/lua_global.cpp
@@ -284,6 +284,8 @@ void LuaReloadActiveMods()
 	LoadItemData();
 	LoadObjectData();
 	LoadQuestData();
+	LoadSetLevelNames();
+	LoadQuestPools();
 	LoadLevelGenerationData();
 
 	lua::LoadModsComplete();